Add SharedTree summary load benchmarks
